fixup: extract allowlist_file submodule; allow dead_code on macOS
- Pull the file-reading helpers (read, parse_line, canonicalise_entry,
  target_safe) out of peer_check/linux.rs into peer_check/allowlist_file.rs
  with their own tests. Drops the cfg(test) gates from those helpers; only
  load_default in linux.rs still has a cfg(test) split per the
  controller-test allowlist override.
- Module-level allow(dead_code) for the parent peer_check on macOS: the
  Swift extension is the production path there, so `verify_peer` always
  returns Unverifiable and the non-Unverifiable PeerRejected variants +
  Allowlist::paths field are never constructed. Suppresses the lint
  rather than gating each variant individually.
